Add incident banners in extension
